% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Input.R
\name{Format_Input}
\alias{Format_Input}
\title{Convert raw data into a \code{data.frame}}
\usage{
Format_Input(id, wave, income, CPI, weight = 1, bottom.threshold = 1,
  top.threshold)
}
\arguments{
\item{id}{Vector of a unique longitudinal identifier for a survey unit across all waves.}

\item{wave}{Vector of survey data collection periods.}

\item{income}{Vector of income corrsponding to \code{id} and \code{wave}.}

\item{CPI}{Vector of CPI corrsponding to \code{wave}.}

\item{weight}{Vector of survey weight corrsponding to \code{id}.}

\item{bottom.threshold}{Vector of income bottom-coding threshold corrsponding to \code{id} and \code{wave}, or a single value.}

\item{top.threshold}{Vector of income top-coding threshold corrsponding to \code{id} and \code{wave}, or a single value.}
}
\value{
A \code{data.frame} contains all nesscary information for futher processing in \link{GB2MLE} and \link{Imputation_Method}.
}
\description{
This function process the raw longitudinal information and store them in a \code{data.frame}
}
\references{
Tan, Li (2017), Imputing Top-Coded Income Data in Longitudinal Surveys, Working Papers. (\href{http://litaneconomics.com/Job_Market_Paper.pdf}{link})
}
