% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Method.R
\name{Imputation_Method}
\alias{Imputation_Method}
\title{The Imputated Income Values Simulator}
\usage{
Imputation_Method(df, par = NULL, standard = 1, rank_based = 1,
  Bayesian = 1, select = NA, prior.grid = 50, rtol = 1e-11)
}
\arguments{
\item{df}{The longitudinal file.}

\item{par}{The GB2 distribution parameters. If not provided, the parameters will be estimated based \code{df}.}

\item{standard}{Flag for generating imputed values by the standard method in the output.}

\item{rank_based}{Flag for generating imputed values by the rank_based method in the output.}

\item{Bayesian}{Flag for generating imputed values by the Bayesian method in the output.}

\item{select}{How many waves are used to estimate prior distribution. If NA, then all waves will be used.}

\item{prior.grid}{Controls the number of bins in \link[REBayes]{WGLVmix}.}

\item{rtol}{The relative tolerance for dual gap convergence criterion used in \link[REBayes]{KWDual}.}
}
\value{
A \link{data.frame} contains the input with extra columns shown the imputed income values.
}
\description{
A function to populate imputed income values by the standard, rank-based and Bayesian methods.
}
\examples{
Imputation_Method(df = SIPP1996Analytic, par = Example_GB2_param, standard = 1, rank_based = 1, Bayesian = 1, select = NA, prior.grid = 50, rtol = 1e-11)

}
\references{
Tan, Li (2017), Imputing Top-Coded Income Data in Longitudinal Surveys, Working Papers. (\href{http://litaneconomics.com/Job_Market_Paper.pdf}{link})
}
